pyproject.toml
[project]
name = "retaico-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.5.2",
    "pandas>=3.0.2",
    "pandera>=0.30.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
